pyproject.toml
[tool.poetry]
name = "private-residential-property-transaction"
version = "0.1.0"
description = ""
authors = ["Nicholas <nictann96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
numpy = "^2.0.0"
requests = "^2.32.3"
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
python-dotenv = "^1.0.1"
pymongo = "^4.8.0"
tabula-py = "^2.9.3"
sqlalchemy = "^2.0.32"
openpyxl = "^3.1.5"
flask = "^3.0.3"
dash = "^2.18.1"
dash-html-components = "^2.0.0"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
